+++ cmd/admin.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2025 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2026 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -49,7 +49,7 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/generic"
 	"github.com/arangodb/kube-arangodb/pkg/util/kclient"
-	"github.com/arangodb/kube-arangodb/pkg/util/token"
+	utilToken "github.com/arangodb/kube-arangodb/pkg/util/token"
 )
 
 const (
@@ -431,13 +431,13 @@ func getJWTTokenFromSecrets(ctx context.Context, secrets generic.ReadClient[*cor
 		return nil, errors.WithMessage(err, fmt.Sprintf("failed to get secret \"%s\"", name))
 	}
 
-	claims := token.NewClaims().With(
-		token.WithDefaultClaims(),
-		token.WithServerID("kube-arangodb"),
+	claims := utilToken.NewClaims().With(
+		utilToken.WithDefaultClaims(),
+		utilToken.WithServerID("kube-arangodb"),
 	)
 
 	if len(paths) > 0 {
-		claims = claims.With(token.WithAllowedPaths(paths...))
+		claims = claims.With(utilToken.WithAllowedPaths(paths...))
 	}
 
 	authz, err := claims.Sign(secret)